creacion rutas post para admin
