--- src/routes/admin/post_clientes.js
@@ -1,9 +1,0 @@
-const { clientes } = require('../../controllers/admin/clientes.js');
-const { Router } = require('express');
-const router = Router();
-
-
-router.post('/clientes', clientes)
-
-
-module.exports = router;
