--- src/routes/admin/post_unidad_medida.js
@@ -1,9 +1,0 @@
-const { um } = require("../../controllers/admin/unidad_medida");
-const { Router } = require("express");
-const router = Router();
-
-// Ruta: http://localhost:3001/admin/um
-
-router.post("/um", um);
-
-module.exports = router;
